requirements.txt
Pillow>=10.0.0
mss>=9.0.0
openai>=1.0.0
python-dotenv>=1.0.0
treys>=0.1.8
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0
